Add connection statistic upgrade
Interval based statistics are now available. You have to provide a
Statistics object to the connection. This is a breaking change.
The data will be used to calculate the interval based statistics.
This allows you to apply your own interval.

Also the statistics object has now 3 sub-object for the accumulated,
interval, and instantaneous values.