fix broken connections in case of high latency
When listening and accepting an incoming connection request, the
response might be received by the peer with some delay due to latency.
This causes the peer to send a second connection request, that is not
detected as duplicate because the first connection request has already
been removed from the map that is used to check for duplicates
(connReqs), so it is treated as a brand new connection request,
breaking the first connection.

This patch fixes the issue by replacing the map that is used to check
for duplicates (connReqs) with a map that keeps track of both
connection requests and accepted connections by their peer ID
(connByPeers), allowing to discard handshakes with an existing peer ID.